create delete link and delete methods

+++ WebContent/list-students.jsp
@@ -47,7 +47,7 @@
 							<td>
 							<a href="${updateLink}">Update</a>
 							| 
-							<a href="${deleteLink}">Delete</a>
+							<a href="${deleteLink}" onclick="if (!(confirm('Are you sure you want to delete this student?'))) return false">Delete</a>
 							</td>
 						</tr>
 					</c:forEach>
